Add test coverage for design doc cache properties
While there isn't a documented formal guarantee for dbcore when
the W=N, there is some expectation that eviction will happen in
a timely manner. This test does some light stressing of this
process to make sure that proper designs are being used for
views and validation.

I've omitted some more aggressive tests that seem to exploit
other possible race conditions to investigate the properties in
more depth later. I'm not sure where to put tests that are
expected to cause rare failures only after hundreds of
iterations.